fix(plan9): properly pack texture structs
The PACKEDATTR macro uses GCC's __attribute__((packed)) which is
obviously not supported by the Plan9 compiler.
The macro defaults to no packing at all, causing problems when loading
the textures since the structs are of different sizes.

I could not find any information about packing structs in plan9 but
poking around the source of the 6c compiler I found a packflg flag and
decided to give it a try:

	#pragma pack on

	// struct definitions

	#pragma pack off

Surprisingly it worked first try!